+++ NpgsqlRest/UploadHandlers/Handlers/LargeObjectUploadHandler.cs
@@ -65,7 +65,7 @@ public class LargeObjectUploadHandler(RetryStrategy? retryStrategy) : BaseUpload
 
         if (Options.UploadOptions.LogUploadParameters is true)
         {
-            Logger?.LogDebug("Upload for {_type}: includedMimeTypePatterns={includedMimeTypePatterns}, excludedMimeTypePatterns={excludedMimeTypePatterns}, bufferSize={bufferSize}, oid={oid}, checkText={checkText}, checkImage={checkImage}, allowedImage={allowedImage}, testBufferSize={testBufferSize}, nonPrintableThreshold={nonPrintableThreshold}", 
+            Logger?.LogTrace("Upload for {_type}: includedMimeTypePatterns={includedMimeTypePatterns}, excludedMimeTypePatterns={excludedMimeTypePatterns}, bufferSize={bufferSize}, oid={oid}, checkText={checkText}, checkImage={checkImage}, allowedImage={allowedImage}, testBufferSize={testBufferSize}, nonPrintableThreshold={nonPrintableThreshold}", 
                 Type, IncludedMimeTypePatterns, ExcludedMimeTypePatterns, BufferSize, oid, checkText, checkImage, allowedImage, testBufferSize, nonPrintableThreshold);
         }